% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{bp_results}
\alias{bp_results}
\title{test_for_hetero}
\usage{
bp_results(model_object)
}
\arguments{
\item{model_object}{from \code{get_model}}
}
\value{
test results
}
\description{
This function will run a test for heteroscedasticity within the data.
}
