% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a-connectivity.R
\name{conn_summary}
\alias{conn_summary}
\title{conn_summary}
\usage{
conn_summary(wrangled_devices)
}
\arguments{
\item{ts}{wrangled_devices : data.frame}
}
\value{
wrangled_devices Data Frame Full Joined with Obs
}
\description{
Connectivity Summary Kable
}
