% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a-connectivity.R
\name{conn_wrangle}
\alias{conn_wrangle}
\title{conn_wrangle}
\usage{
conn_wrangle(wrangled_devices, filtered_obs)
}
\arguments{
\item{ts}{wrangled_obs     : data.frame obs table}
}
\value{
wrangled_devices Data Frame Joined with Obs
}
\description{
Connectivity Wrangle - Primary table is devices : not obs
}
