% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{data_segments}
\alias{data_segments}
\title{data_segments}
\usage{
data_segments(wrangled_devices, target_var)
}
\arguments{
\item{wrangled_devices}{data.frame, as output by \code{wrangle_devices}}

\item{target_var}{list, report parameters which we will segment upon}
}
\description{
work out the max number of data segments - useful for formatting plots
}
