% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{date_string}
\alias{date_string}
\title{date_string}
\usage{
date_string(ts)
}
\arguments{
\item{ts}{timestamp}
}
\value{
Date in {mmm dd, YYYY} format
}
\description{
Create nicely formatted sgtring from ts date
}
