% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{device_excluded_kable}
\alias{device_excluded_kable}
\title{excluded_device_kable}
\usage{
device_excluded_kable(data_volume)
}
\arguments{
\item{ts}{wrangled_devices : data.frame device table!}
}
\value{
kable
}
\description{
Summary Table of Devices Excluded
}
