% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{device_summary_string}
\alias{device_summary_string}
\title{device_summary_string}
\usage{
device_summary_string(devices_for_map)
}
\arguments{
\item{ts}{devices_for_map : data.frame with device connection details}
}
\value{
Character[] The text for Device Summary Section
}
\description{
Breif Summary of Device Connection Rate etc
}
