% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{downsample_obs}
\alias{downsample_obs}
\title{downsample_obs}
\usage{
downsample_obs(wrangle_obs, rate)
}
\arguments{
\item{wrangle_obs}{from \code{wrangle_observations}}

\item{rate}{downsampling rate}
}
\description{
Reduce data set for intensive charts (plotly mostly)
}
\examples{
wrangle_obs <- downsample_obs(wrangle_obs, 5)

}
