% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{extract_group_target_vars}
\alias{extract_group_target_vars}
\title{extract_group_target_vars}
\usage{
extract_group_target_vars(wrangled_devices, param_list)
}
\arguments{
\item{wrangled_devices}{data.frame, as output by \code{wrangle_devices}}

\item{param_list}{list, report parameters from param_list. The ones we want will take the form
'Analyse {param} group' - for instance 'Analyse Energymate group'}
}
\description{
adds new columns to segment data by groups (if present in data)
}
