% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{filter_list}
\alias{filter_list}
\title{filter_list}
\usage{
filter_list(available, permitted)
}
\arguments{
\item{available}{list, readings we have in our dataset}

\item{permitted}{list, list of permitted readings types}
}
\description{
returns the valid readings. Of those available, it filters against
the list of permitted values - stops propogation of unexpected values.
}
