% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{geojson_verify_colour_col}
\alias{geojson_verify_colour_col}
\title{geojson_verify_colour_col}
\usage{
geojson_verify_colour_col(geojson, map_params, map_layer)
}
\arguments{
\item{geojson_file}{sf}
}
\value{
sf : geojson map object as sf
}
\description{
Generate missing Colour Fill Col for geojson object
}
\examples{
geojson_main  <- geojson_verify_colour_col(
  geojson = geojson_main,
  map_params = map_params,
  map_layer = "main"
)

}
