% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_clean_model_data}
\alias{get_clean_model_data}
\title{get_clean_model_data}
\usage{
get_clean_model_data(wrangled_obs_interv, model_vars, output_var)
}
\arguments{
\item{wrangled_obs_interv}{from \code{wrangle_interv}}

\item{model_vars}{a vector of string input variables/predictors by name}

\item{output_var}{a string indicating the column name of the output/predicted variable}
}
\value{
List
}
\description{
We need clean data for the model. This function will join the joined obs
and interv data with the device and weather data. After those files are all joined, the
function will then clean the data and keep only the data-points we need. A list of
two items will be returned. The Output Variable, and the cleaned data
}
