% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-wrangle.R
\name{get_data_volume}
\alias{get_data_volume}
\title{get_data_volume}
\usage{
get_data_volume(observations, from_timestamp, to_timestamp)
}
\arguments{
\item{observations}{from \code{wrangle_observations}}

\item{from_timetamp, to_timestamp}{timestamps for beginning & end of the report}
}
\value{
compact data frame with : \cr
\itemize{
\item device_id,  \cr
\item n() records,  \cr
\item r \% expected,  \cr
\item exclude. \cr
}
}
\description{
Check data volume and flag to exclude devices with < 10\% data
}
\examples{
data_volume <- get_data_volume(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
  )
}
