% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_exposure_kable_dev}
\alias{get_exposure_kable_dev}
\title{get_exposure_kable_dev}
\usage{
get_exposure_kable_dev(temp_exposure, target_var)
}
\arguments{
\item{temp_exposure}{from \code{get_temp_exposure}}

\item{target_var}{variable upon which to segment kable}
}
\value{
Kable
}
\description{
Using temp_exposure data-set, generate a neat Kable showing
temperature exposure for all devices individually.
}
\examples{
temp_exposure      <- get_temp_exposure(wrangled_obs, wrangled_devices, target_var)
exposure_kable_var <- get_exposure_kable_var(temp_exposure, target_var)
exposure_kable_dev <- get_exposure_kable_dev(temp_exposure, target_var)

}
