% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_exposure_plot}
\alias{get_exposure_plot}
\title{get_exposure_plot}
\usage{
get_exposure_plot(exposure_chart_data, target_var, chart_type = "bar")
}
\arguments{
\item{exposure_chart_data}{from \code{get_exposure_chart_data} for example}

\item{target_var}{variable upon which to segment kable}

\item{chart_type}{"bar" or "donut" - defaults to "bar"}
}
\value{
Data.Frame
}
\description{
Takes well formatted data and outputs exposure plot against
WHO thresholds - is currently used with :
\itemize{
\item temperature exposure analysis
\item lowest temperature experienced
}
}
\examples{
exposure_chart_data <- get_exposure_chart_data(wrangled_obs, wrangled_devices, target_var)
exposure_plot <- get_exposure_plot(exposure_chart_data, target_var)

lowtemp_chart_data <- get_lowtemp_chart_data(wrangled_obs, wrangled_devices, target_var)
lowtemp_chart      <- get_exposure_plot(lowtemp_chart_data, target_var)

}
