% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1e-home-report-HCHO.R
\name{get_hcho_totals}
\alias{get_hcho_totals}
\title{get_hcho_totals}
\usage{
get_hcho_totals(hcho_data)
}
\arguments{
\item{hcho_data}{data.frame, the output from \code{?hcho_data()}}
}
\description{
Totals for Value Boxes
}
