% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-home-report-WHO.R
\name{get_heat_palette}
\alias{get_heat_palette}
\title{get_heat_palette}
\usage{
get_heat_palette(width)
}
\arguments{
\item{width}{number, the number of hex values needed}
}
\description{
Get Heat Palette Colour Hex Values
}
