% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_hetero_model}
\alias{get_hetero_model}
\title{get_hetero_model}
\usage{
get_hetero_model(clean_model_data, model_object, output)
}
\arguments{
\item{clean_model_data}{from \code{get_clean_model_data}}

\item{model_object}{from \code{get_model}}

\item{output_var}{from \code{get_clean_model_data}}
}
\value{
model_object
}
\description{
This function will remake the model, with weights that will help fix the heteroscedasticity.
}
