% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_intervention_sample}
\alias{get_intervention_sample}
\title{Get Intervention Sample}
\usage{
get_intervention_sample(
  wrangled_obs_interv,
  intervention_data,
  interventions_only,
  hours_greater_than,
  hours_less_than,
  temps_greater_than,
  temps_less_than
)
}
\arguments{
\item{wrangled_obs_interv}{from \code{wrangle_interv}}

\item{interventions_only}{a logical argument about whether observations should be restricted to homes with interventions}

\item{hours_greater_than}{an integer indicating night time start}

\item{hours_less_than}{an integer indicating night time end}

\item{temps_greater_than}{an integer indicating the lowest outdoor temp}

\item{temps_less_than}{an integer indicating the highest outdoor temp}

\item{inteverntion_data}{a data frame with all the interventions to be evaluated}
}
\value{
a data frame of filtered observations
}
\description{
A function to filter observations based on several parameters
}
