% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{get_local_time}
\alias{get_local_time}
\title{Unix timestamp to local timezone datetime}
\usage{
get_local_time(timestamp, lat, long)
}
\arguments{
\item{timestamp}{Unix time stamp vector}

\item{lat}{Latitude as vector}

\item{long}{Longitude as vector}
}
\value{
Returns "POSIXct", "POSIXt" vector in local time zone
}
\description{
Convert time stamp to local timezone datetime
}
\examples{
get_local_time(1555555555, -125.357, 23.556)

}
