% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{get_local_time_fast}
\alias{get_local_time_fast}
\title{get_local_time_fast}
\usage{
get_local_time_fast(timestamp, local_tz)
}
\arguments{
\item{timestamp}{Unix time stamp vector}

\item{local_tz}{local timezone}
}
\value{
Returns "POSIXct", "POSIXt" vector in local time zone
}
\description{
Convert time stamp to local timezone datetime
}
\examples{
get_local_time(1555555555, -125.357, 23.556)

}
