% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{get_local_tz}
\alias{get_local_tz}
\title{get_local_tz}
\usage{
get_local_tz(wrangled_devices)
}
\arguments{
\item{wrangled_devices}{}
}
\value{
local_tz
}
\description{
Based on the wrangled devices :\cr
\itemize{
\item Exclude NZ or missing lon / lat \cr
\item Grab average values \cr
\item Used for time-series axis \cr
}
}
\examples{
get_local_tz(wrangled_devices = wrangled_devices)

}
