% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_lowest_ever_kable_var}
\alias{get_lowest_ever_kable_var}
\title{get_lowest_ever_kable_var}
\usage{
get_lowest_ever_kable_var(lowest_ever, target_var)
}
\arguments{
\item{lowest_ever}{from \code{get_temp_lowest_ever}}

\item{target_var}{variable upon which to segment kable}
}
\value{
Kable
}
\description{
Using lowest_ever data-set, generate a neat Kable showing
temperature exposure groups by a target var.
}
\examples{
lowest_ever           <- get_temp_lowest_ever(wrangled_obs, wrangled_devices, target_var)
lowest_ever_kable_var <- get_lowest_ever_kable_var(lowest_ever, target_var)

}
