% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_lowtemp_chart_data}
\alias{get_lowtemp_chart_data}
\title{get_lowtemp_chart_data}
\usage{
get_lowtemp_chart_data(wrangled_obs, wrangled_devices, target_var)
}
\arguments{
\item{wrangled_obs}{from \code{wrangled_obs}}

\item{wrangled_devices}{from \code{wrangled_devices}}

\item{target_var}{variable upon which to segment kable}
}
\value{
Data.Frame
}
\description{
Starting from observations wrangled for use testing against WHO
thresholds, this function massages the data into the requisiste format for the
get_exposure_plot function.
}
\examples{
lowtemp_chart_data <- get_lowtemp_chart_data(wrangled_obs, wrangled_devices, target_var)
lowtemp_chart      <- get_exposure_plot(lowtemp_chart_data, target_var)

}
