% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{get_map}
\alias{get_map}
\title{get_map}
\usage{
get_map(devices_for_map, map_params, geojson_main, geojson_inset)
}
\arguments{
\item{devices_for_map}{data.frame : from \code{monkeyr::get_devices_for_map()}}

\item{map_params}{list :}

\item{geojson_main}{MAP :}

\item{geojson_inset}{MAP :}
}
\value{
plot
}
\description{
generate map with inset
}
\examples{
map  <- update_map_params(
  devices_for_map = devices_for_map,
  map_params = map_params
)

}
