% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{get_map_params}
\alias{get_map_params}
\title{get_map_params}
\usage{
get_map_params(countryCode_local)
}
\arguments{
\item{countryCode}{string : the country code ("NZ", "AU", ...)}
}
\value{
list : country map parameters
}
\description{
Get map params for target country as list
}
\examples{
map_params  <- get_map_params("NZ")

}
