% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_model}
\alias{get_model}
\title{get_model}
\usage{
get_model(clean_data_list)
}
\arguments{
\item{clean_data}{from `get_clean_model_data}
}
\value{
model_object
}
\description{
This function will get the model object so we can get items like the coefficients and
residuals.
}
