% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_model_coef}
\alias{get_model_coef}
\title{get_model_coef}
\usage{
get_model_coef(model)
}
\arguments{
\item{model_object}{from \code{get_model}}
}
\value{
Data.Frame
}
\description{
This function will get the model object and return the coefficients of the model
}
