% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_model_residuals}
\alias{get_model_residuals}
\title{get_model_residuals}
\usage{
get_model_residuals(model, clean_data)
}
\arguments{
\item{clean_data}{from \code{get_clean_model_data}}

\item{model_object}{from \code{get_model}}
}
\value{
Data.Frame
}
\description{
This function will get the model object and return the output variable indoor temp
and the residuals in a data frame
}
