% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_model_rsquared}
\alias{get_model_rsquared}
\title{get_model_rsquared}
\usage{
get_model_rsquared(model)
}
\arguments{
\item{model_object}{from \code{get_model}}
}
\value{
list
}
\description{
This function will get the model object and return the r-squared and adjusted
r-squared as a list.
}
