% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{get_palette}
\alias{get_palette}
\title{get_palette}
\usage{
get_palette(reading, palette_depth)
}
\arguments{
\item{reading}{string, which reading: {TEMP, HUM ...}}

\item{palette_depth}{number, how many colours?}
}
\description{
get a single palette, set per reading type
}
