% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d-home-report-PM.R
\name{get_pm_totals}
\alias{get_pm_totals}
\title{get_pm_totals}
\usage{
get_pm_totals(pm_data)
}
\arguments{
\item{pm_data}{data.frame, the output from \code{?pm_data()}}
}
\description{
Totals for Value Boxes
}
