% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_reg_data}
\alias{get_reg_data}
\title{get_reg_data}
\usage{
get_reg_data(wrangled_devices, wrangled_obs, target_reading = "temp")
}
\arguments{
\item{wrangled_devices}{from \code{wrangle_observations( )}}

\item{wrangled_obs}{from \code{wrangled_obs( )}}

\item{target_reading}{variable upon which to segment data}
}
\value{
data frame
}
\description{
Make Data-Frame of Regression Test:
for use with indoor - outdoor regression testing
}
\examples{
reg_data  <- get_reg_data(wrangled_devices, wrangled_obs, target_reading = "temp")
reg_kable <- get_reg_kable(reg_data, target_var)

}
