% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_reg_kable}
\alias{get_reg_kable}
\title{get_reg_kable}
\usage{
get_reg_kable(reg_data, target_var)
}
\arguments{
\item{reg_data}{from \code{get_reg_data}}

\item{target_var}{variable upon which to segment kable}
}
\value{
kable
}
\description{
Make Kable from Inside-Outside Regression Data that shows detailed
device's model relationship coefficient.
}
\examples{
reg_data  <- get_reg_data(wrangled_devices, wrangled_obs, target_reading = "temp")
reg_kable <- get_reg_kable(reg_data, target_var)

}
