% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_temp_exposure}
\alias{get_temp_exposure}
\title{get_temp_exposure}
\usage{
get_temp_exposure(wrangled_obs, wrangled_devices, target_var)
}
\arguments{
\item{wrangled_obs}{from \code{wrangled_obs}}

\item{wrangled_devices}{from \code{wrangled_devices}}

\item{target_var}{variable upon which to segment kable}
}
\value{
Data.Frame
}
\description{
Generate Temperature Exposure data-set used to build kables that
show percentage of time spent at night below WHO thresholds.
}
\examples{
temp_exposure      <- get_temp_exposure(wrangled_obs, wrangled_devices, target_var)
exposure_kable_var <- get_exposure_kable_var(temp_exposure, target_var)
exposure_kable_dev <- get_exposure_kable_dev(temp_exposure, target_var)

}
