% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-temp_analysis.R
\name{get_temp_lowest_ever}
\alias{get_temp_lowest_ever}
\title{get_temp_lowest_ever}
\usage{
get_temp_lowest_ever(wrangled_obs, wrangled_devices, target_var)
}
\arguments{
\item{wrangled_obs}{from \code{wrangled_obs}}

\item{wrangled_devices}{from \code{wrangled_devices}}

\item{target_var}{variable upon which to segment kable}
}
\value{
Data.Frame
}
\description{
Generate Lowest Ever Temperature data-set used to build kables
showing percentage of devices which dipped below thresholds.
}
\examples{
lowest_ever           <- get_temp_lowest_ever(wrangled_obs, wrangled_devices, target_var)
lowest_ever_kable_var <- get_lowest_ever_kable_var(lowest_ever, target_var)

}
