% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c-home-report-Ventilation.R
\name{get_ventilation_totals}
\alias{get_ventilation_totals}
\title{#' get_heat_palette
#' @description Get Heat Palette Colour Hex Values
#' @param width number, the number of hex values needed
#' @export
get_heat_palette <- function(width) {
tryCatch (
{
heatcols <- 6
switch(width,
{heatcols <- c("#F97162")},
{heatcols <- c("#F97162", "#FBE1B1")},
{heatcols <- c("#F97162", "#FBE1B1", "#B7DFCB")},
{heatcols <- c("#F97162", "#FBE1B1", "#B7DFCB", "#5ABAD1")},
{heatcols <- c("#F97162", "#FBE1B1", "#B7DFCB", "#5ABAD1", "#3984B6")},
{heatcols <- c("#F97162", "#FBE1B1", "#B7DFCB", "#5ABAD1", "#3984B6", "#275a7c")},
)}
\usage{
get_ventilation_totals(ventilation_data)
}
\arguments{
\item{who_threshold_data}{data.frame, the output from \code{?ventilation_data()}}
}
\description{
Totals for Value Boxes
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## Error Handler
\},
  error = function(cond) \{
    monkeyr::monkey_knit_error(err = cond, resource = "get_heat_palette")
\}
}\if{html}{\out{</div>}}

)
}

#' degrees_below_18_score
#' @description Calculate the degrees below 18C score
#' @param observations data.frame, the wrangled observations data frame as output by
#' \code{wrangle.observations} and further filtered by \code{remove_excluded_devices}
#' @export
degrees_below_18_score <- function(observations) {
tryCatch (
{
## average degrees below 18
browser()
score <- observations \%>\%
## only intereted in 11pm - 7am
dplyr::filter(as.numeric(hour) \%in\% c(23, 0:6)) \%>\%
dplyr::mutate(cap_co2 = ifelse(co2 > 18, 18, co2)) \%>\%
dplyr::summarise(sum(18 - cap_co2)) / nrow(.)

\if{html}{\out{<div class="sourceCode">}}\preformatted{  ## shitness score. 12deg (score = 6) -> 100.
  score <- 100 - 100 * (score / (18 - 12))
  floor(max(score, 0))

## Error Handler
\},
  error = function(cond) \{
    monkeyr::monkey_knit_error(err = cond, resource = "degrees_below_18_score")
\}
}\if{html}{\out{</div>}}

)

}

get_ventilation_totals
}
