% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1e-home-report-HCHO.R
\name{hcho_data}
\alias{hcho_data}
\title{hcho_data}
\usage{
hcho_data(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Use WHO guidelines functions on Formaldehyde
}
