% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1e-home-report-HCHO.R
\name{hcho_guage}
\alias{hcho_guage}
\title{hcho_guage}
\usage{
hcho_guage(hcho_data, from_timestamp, to_timestamp)
}
\arguments{
\item{hcho_data}{data.frame, the output from \code{?hcho_data()}}

\item{from_timestamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
Guage Charts facetted per device
}
