% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{hhi_list_kable}
\alias{hhi_list_kable}
\title{hhi_list_kable}
\usage{
hhi_list_kable(devices_for_map, cen_top_level)
}
\arguments{
\item{devices_for_map}{from \code{get_devices_for_map()}}

\item{cen_top_level}{(Boolean) is this a CEN top level report}
}
\value{
kable & description line
}
\description{
Summary Table of All HHIs (CEN Report only)
}
