% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{home_summary_kable}
\alias{home_summary_kable}
\title{home_summary_kable}
\usage{
home_summary_kable(wrangled_devices)
}
\arguments{
\item{ts}{wrangled_devices : data.frame device table!}
}
\value{
kable
}
\description{
Summary Table By Address
}
