% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit-home-rmd.R
\name{knit_home_report}
\alias{knit_home_report}
\title{Compile the home report}
\usage{
knit_home_report(
  param_list,
  output_file = paste0("home-", as.numeric(Sys.time()), ".html"),
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{param_list}{named list of parameters, see \code{?make_param_list}}

\item{output_file}{name of the output html (default: home-{timestamp}.html)}

\item{output_dir}{path to folder to save the html (default: current work dir)}

\item{...}{additional named arguments passed onto \code{rmarkdown::render}}
}
\value{
a compiled document written into the output file and the path of the output file is returned
}
\description{
Create a home report with a given set of pameters.
}
