% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{knit_report}
\alias{knit_report}
\title{Compile the home report}
\usage{
knit_report(
  param_list,
  output_file = paste0("home-", as.numeric(Sys.time()), ".html"),
  output_dir = getwd(),
  rmd_file = "home.rmd",
  ...
)
}
\arguments{
\item{param_list}{named list of parameters, see \code{?make_param_list}}

\item{output_file}{name of the output html (default: home-{timestamp}.html)}

\item{output_dir}{path to folder to save the html (default: current work dir)}

\item{...}{additional named arguments passed onto \code{rmarkdown::render}}
}
\value{
a compiled document written into the output file and the path of the output file is returned
}
\description{
Create a home report with a given set of pameters.
}
\examples{
obs  <- system.file("data/filtered_obs.csv", package="monkeyr")
dev  <- system.file("data/filtered_devices.csv", package="monkeyr")
wea  <- system.file("data/filtered_weather.csv", package="monkeyr")
int  <- system.file("data/filtered_interventions.csv", package="monkeyr")
from <- 1627639151
to   <- 1630317551

par_list <- make_params_list(
  filtered_obs = obs,
  filtered_devices = dev,
  filtered_weather = weather,
  filtered_interventions = interv,
  params = report_params,
  fromTimeStamp = from,
  toTimeStamp = to
) {
knit_report(par_list)

}

}
