% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{make_nz_datetime}
\alias{make_nz_datetime}
\title{Unix timestamp to local timezone datetime}
\usage{
make_nz_datetime(timestamp, lat, long)
}
\arguments{
\item{timestamp}{Unix time stamp vector}

\item{lat}{Latitude! as vector}

\item{long}{Longitude! as vector}
}
\value{
Returns "POSIXct", "POSIXt" vector in NZ time zone
}
\description{
Convert a time stamp into local timezone datetime
}
\examples{
make_nz_datetime(1555555555, -125.357, 23.556)
}
