% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{make_params_list}
\alias{make_params_list}
\title{Make a list of parameters for the report}
\usage{
make_params_list(.list = NULL, ...)
}
\arguments{
\item{.list}{A named list of key-value pairs}

\item{...}{Unquoted key-value pairs}
}
\value{
A list of RMD parameters that may be passed to a report generating function
}
\description{
Create a list of parameters for building the report. Params determine the data and time
period used in each report. Review the details below for a list of fields to pass as parameters
}
\details{
Field options are:\cr
\itemize{
\item filtered_obs (a dataframe or filepath to a csv),
\item filtered_weather (a dataframe or filepath to a csv),
\item filtered_devices (a dataframe or filepath to a csv),
\item filtered_interventions (a dataframe or filepath to a csv),
\item report_params,
\item fromTimeStamp,
\item toTimeStamp
}
}
\examples{
# convenient for interactive typing
make_params_list(filtered_obs = "inst/data/filtered_obs.csv")
# convenient for programming
make_params_list(.list = list(filtered_obs = "inst/data/filtered_obs.csv"))

}
