% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{make_polygon_area}
\alias{make_polygon_area}
\title{make_polygon_area}
\usage{
make_polygon_area(observations, target_variable, severity)
}
\arguments{
\item{observations}{data.frame, the output of \code{wrange_observations}}

\item{target_variable}{This can be {"hum", "temp", "co2" ...}}

\item{severity}{This can be {1, 2, 3} which band it defines.}
}
\value{
a tibble dataframe of polygon data
}
\description{
Make little coloured area in plot
}
