% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-test-params.R
\name{make_test_params}
\alias{make_test_params}
\title{Create a List of Parameters to Test}
\usage{
make_test_params(data_folder = "monkey_a", params_file = "params_1.txt", ...)
}
\arguments{
\item{params_file}{params txt file located at inst/extdata/params/{params_file}}

\item{folder}{folder located in inst/extdata/{folder}}
}
\value{
a list of parameters that may be used for generating a report
}
\description{
Test Data! Data-sets include obs.csv; devices.csv;
weather.csv & intreventions.csv. Plus report params as txt file.
We will collect data-sets from real world tests as we go.
Example data-sets might be : \cr
\itemize{
\item All devices in data-set have no data.
\item Only 1-2 devices in data-set. Mimic home user.
\item Lots of devices. Lots of data. Mimic DHB data.
\item etc... Expand the cases as they arise.
}
}
