% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{memory_saver_obs}
\alias{memory_saver_obs}
\title{memory_saver_obs}
\usage{
memory_saver_obs(wrangled_obs)
}
\arguments{
\item{wrangle_obs}{from \code{wrangle_observations}}
}
\description{
Cut out unnecessary data to reduce memory by factor of 3-4
}
\examples{
wrangle_obs <- memory_saver_obs(wrangle_obs)

}
