% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{monkey_knit_error}
\alias{monkey_knit_error}
\title{monkey_error}
\usage{
monkey_knit_error(err = "err message", resource = "failing r function", ...)
}
\arguments{
\item{folder}{folder located in inst/extdata/{folder}}

\item{params_file}{params txt file located at inst/extdata/params/{params_file}}
}
\value{
a list of parameters that may be used for generating a report
}
\description{
R Code crashed. Need to print out message to: \cr
\itemize{
\item to html report
\item console / terminal
\item render
}
}
