% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{monkey_knit_msg}
\alias{monkey_knit_msg}
\title{monkey_knit_msg}
\usage{
monkey_knit_msg(
  msg = "message",
  resource = "failing r function",
  debug = FALSE,
  ...
)
}
\arguments{
\item{msg}{string to holler out everywhere}

\item{resource}{the function or chunk name where it's called}
}
\description{
Debug : print out message to: \cr
\itemize{
\item to html report
\item console / terminal
\item render
}
}
\examples{
monkey_knit_msg(msg = "Thing to output everywhere", resource="CHUNK_NAME")
monkey_knit_msg(msg = "Debugging stuff locally", resource="function name etc.", debug = TRUE)

}
