% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{org_string}
\alias{org_string}
\title{org_string}
\usage{
org_string(wrangled_devices)
}
\arguments{
\item{ts}{data_volume : data.frame with device connection details}
}
\value{
Character[] The text for Device Summary Section
}
\description{
Brief Summary of Device Success / Failure
}
