% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{param_list}
\alias{param_list}
\title{param_list}
\usage{
param_list(report_params, split_char = ",")
}
\arguments{
\item{report_params}{The string with '-' separation}
}
\description{
returns report params as a list
}
