% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d-home-report-PM.R
\name{pm_chart}
\alias{pm_chart}
\title{pm_chart}
\usage{
pm_chart(pm_data, from_timestamp, to_timestamp)
}
\arguments{
\item{pm_data}{data.frame, the output from \code{?pm_data()}}

\item{from_timestamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
chart based on PM
}
