% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d-home-report-PM.R
\name{pm_data}
\alias{pm_data}
\title{pm_data}
\usage{
pm_data(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Use WHO guidelines functions on PM
}
