% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d-home-report-PM.R
\name{pm_donut}
\alias{pm_donut}
\title{pm_donut}
\usage{
pm_donut(pm_data, from_timestamp, to_timestamp)
}
\arguments{
\item{pm_data}{data.frame, the output from \code{?pm_data()}}

\item{from_timestamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
Donut Chart (facetted per device)
}
