% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{read_geojson}
\alias{read_geojson}
\title{read_geojson}
\usage{
read_geojson(geojson_file)
}
\arguments{
\item{geojson_file}{string : path from \code{map_params$xxxx.geojson}}
}
\value{
sf : geojson map object as sf
}
\description{
Get geojson object from file
}
\examples{
geojson_main  <- read_geojson(
  geojson_file = map_params$main.geojson
)

geojson_inset  <- read_geojson(
  geojson_file = map_params$inset.geojson
)

}
