% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{report_period}
\alias{report_period}
\title{Print a formatted from-to sentence}
\usage{
report_period(from_timestamp, to_timestamp)
}
\arguments{
\item{from_timetamp, to_timestamp}{ts - begin and end of the report.}
}
\value{
Character string
}
\description{
Takes 2 time stamps and prints a formatted from-to sentence.
}
\examples{
report_period(1555555555, 1666666666)

}
