% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{room_summary_kable}
\alias{room_summary_kable}
\title{room_summary_kable}
\usage{
room_summary_kable(wrangled_devices)
}
\arguments{
\item{ts}{wrangled_devices : data.frame device table!}
}
\value{
kable
}
\description{
Summary Table By Room Type (Bedrooms vs Living)
}
