% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit-home-rmd.R
\name{run_test_report}
\alias{run_test_report}
\title{run_test_report}
\usage{
run_test_report(
  report = "home",
  dummy_data_dir = "monkey_a",
  dummy_params_file = "params_1.txt"
)
}
\arguments{
\item{report}{The report name}

\item{dummy_data_dir}{The Dummy data set which can be considered as a test case}

\item{dummy_params_file}{The Dummy parameter set - rmd params into report}
}
\description{
Test Report Locally with Dummy Test Data
and Dummy Test Parameters. Uses library of test cases in inst.
}
\examples{
# run_test_report("home", "monkey_a", "params_1.txt")
# run_test_report("flexi", "org_full", "map-room.txt")
# run_test_report("full", "org_corner1", "map-tenure.txt")
# run_test_report("hhi", "org_corner2", "map.txt")
}
