% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{set_options}
\alias{set_options}
\title{set_options : Need to update this to work with tidy data structure}
\usage{
set_options(wrangled_obs)
}
\arguments{
\item{wrangled_obs}{data.frame, the wrangled observations data frame as output by
\code{wrangle_observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Set options for downstream charts based on available data
}
