% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{summary_kable_by_var}
\alias{summary_kable_by_var}
\title{summary_kable_by_var}
\usage{
summary_kable_by_var(
  wrangled_devices,
  target_var,
  target_var_string,
  unique_owners
)
}
\arguments{
\item{ts}{wrangled_devices : data.frame device table!}
}
\value{
kable
}
\description{
Summary Table By Target Var
}
