% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{target_group_list}
\alias{target_group_list}
\title{target_group_list}
\usage{
target_group_list(wrangled_devices, param_list)
}
\arguments{
\item{wrangled_devices}{data.frame, as output by \code{wrangle_devices}}

\item{param_list}{list, report parameters from param_list. The ones we want will take the form
'{param} anlys' - for instance 'room_type anlys'}
}
\description{
captures requested report group params : to compare to device group as reporting segment
}
