% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{target_var_list}
\alias{target_var_list}
\title{target_var_list}
\usage{
target_var_list(wrangled_obs, param_list)
}
\arguments{
\item{wrangled_obs}{data.frame, as output by \code{wrangle_observations}}

\item{param_list}{list, report parameters from param_list}
}
\description{
matches requested report params to available segments for reporting
}
