% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b-home-report-t&h.R
\name{ts_chart}
\alias{ts_chart}
\title{ts_chart}
\usage{
ts_chart(observations, from_timestamp, to_timestamp, target_variable)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}

\item{target_variable}{character, the name of the column to use for plotting
Options are (for now): \code{temp}, \code{hum}}

\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
make a ggplotly time series chart for temperature, CO2, ...
}
\examples{
ts_chart(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  target_variable = "hum"
)
}
