% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{unique_owners}
\alias{unique_owners}
\title{unique_owners}
\usage{
unique_owners(wrangled_devices)
}
\arguments{
\item{wrangled_devices}{data.frame, as output by \code{wrangle_devices}}
}
\description{
returns the unique device owners - usually only one
}
