% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{unique_readings}
\alias{unique_readings}
\title{unique_readings}
\usage{
unique_readings(wrangled_obs)
}
\arguments{
\item{wrangled_obs}{data.frame, as output by \code{wrangle_observations}}
}
\description{
returns the unique readings available in dataset
}
