% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{update_devices_for_map}
\alias{update_devices_for_map}
\title{update_devices_for_map}
\usage{
update_devices_for_map(
  devices_for_map,
  map_geojson,
  admin_distr_col,
  map_layer
)
}
\arguments{
\item{devices_for_map}{data.frame : from \code{monkeyr::get_devices_for_map()}}

\item{map_geojson}{sf : geojson map object}

\item{admin_distr_col}{string : from \code{map_params$xxxx.admin_distr_col}}

\item{map_layer}{string : either "main" or "inset"}
}
\value{
data.frame
}
\description{
:
Add admin_distr to map object for this map
}
\examples{
devices_for_map  <- update_devices_for_map(
  devices_for_map = devices_for_map,
  map_geojson     = geojson_main,
  admin_distr_col = map_params$main.admin_distr_col,
  map_layer       = "main"
)

}
