% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d-map.R
\name{update_map_params}
\alias{update_map_params}
\title{update_map_params}
\usage{
update_map_params(devices_for_map, map_params)
}
\arguments{
\item{devices_for_map}{data.frame : from \code{monkeyr::get_devices_for_map()}}

\item{map_params}{list :}
}
\value{
list : map_params
}
\description{
Add mode values to map_params
}
\examples{
update_map_params  <- update_map_params(
  devices_for_map = devices_for_map,
  map_params = map_params
)

}
