% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c-home-report-Ventilation.R
\name{ventilation_chart}
\alias{ventilation_chart}
\title{ventilation_chart}
\usage{
ventilation_chart(ventilation_data, from_timestamp, to_timestamp)
}
\arguments{
\item{ventilation_data}{data.frame, the output from \code{?ventilation_data()}}

\item{from_timestamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
Ventilation chart based on CO2
}
