% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c-home-report-Ventilation.R
\name{ventilation_data}
\alias{ventilation_data}
\title{ventilation_data}
\usage{
ventilation_data(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
equivalent to WHO guidalines functions but looking at CO2
}
