% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{wrangle_devices}
\alias{wrangle_devices}
\title{wrangle_devices}
\usage{
wrangle_devices(
  filtered_devices,
  data_type_filter = c("CD", "TH", "PM", "FO", "CL"),
  na.strings = c("", NA),
  integer_columns = c("floor"),
  double_columns = c("long", "lat"),
  character_columns = c("device_owner", "device_id", "data_type", "comms_type",
    "locationCode", "addr", "name", "suburb", "city", "country", "room", "hhi",
    "cluster", "privacy", "risk", "buildingType", "tenure", "height", "direction",
    "school", "building", "classRoom", "deployment_id")
)
}
\arguments{
\item{filtered_devices}{A data.frame or path to file}

\item{data_type_filter}{List of date_types to include}

\item{na.strings}{List of things to ignore, for fread().}

\item{integer_columns}{Columns with integer type.}

\item{double_columns}{Columns with double type.}

\item{character_columns}{Columns with character type.}
}
\value{
Devices Data Frame
}
\description{
Wrangle devices data.\cr
\itemize{
\item Replacing missing parameters with 'unknown' \cr
\item Create room_type { living room / bedroom } \cr
\item Fill in City and Country with best guess (to join with weather) \cr
}
}
\examples{
wrangle_devices(test_params("monkey_a")$dev)

}
