% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{wrangle_interv}
\alias{wrangle_interv}
\title{wrangle_interventions and join with obs}
\usage{
wrangle_interv(filtered_interv, obs)
}
\arguments{
\item{filtered_interv}{data.frame or path to file}

\item{obs}{Wrangled observations df as output by \code{wrangle_obs}}
}
\value{
Observation Data Frame Joined with Weather & Devices & intervention data
}
\description{
Wrangle intervention data\cr
\itemize{
\item Expects obs df (joined with devices & weather) \cr
\item To show the effect of an intervention, only monitors which ultimately receive
that intervention are included - over a fixed time period. \cr
\itemize{
\item data pre-interv 'draft-proof' flagged 'pre-draft-proof' \cr
\item data post-interv 'draft-proof' flagged 'post-draft-proof' \cr
}
\item Comparing devices with an intervention to those without is apples & oranges. \cr
\item Ideally, we want a volume of data: half pre ; half post... \cr
}
}
\examples{
obs <- wrangle_interv(
  test_params("monkey_a")$interv,
  obs = obs
)
}
