% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{wrangle_observations}
\alias{wrangle_observations}
\title{wrangle_observations}
\usage{
wrangle_observations(filtered_obs, devices, weather)
}
\arguments{
\item{filtered_obs}{data.frame or path to file}

\item{devices}{Wrangled devices df as output by \code{wrangle_devices}}

\item{weather}{Wrangled outdoor weather df as output by \code{wrangle_weather}}
}
\value{
Obs Data Frame Joined with Weather & Devices
}
\description{
Wrangle measurement data\cr
\itemize{
\item Filter & dedup (not actually necessary) \cr
\item Join with devices - remove addr & email \cr
\item Fill in missing lat / long - Gen local time\cr
\item Add hour, date, month, year (to join with weather) \cr
}
}
\examples{
wrangled_devices <- wrangle_devices(test_params("monkey_a")$dev)
wrangled_weather <- wrangle_weather(test_params("monkey_a")$weather)
wrangle_observations(
  filtered_obs = test_params("monkey_a")$obs,
  devices = wrangled_devices,
  weather = wrangled_weather
)

}
