% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{wrangle_weather}
\alias{wrangle_weather}
\title{wrangle_weather}
\usage{
wrangle_weather(filtered_weather)
}
\arguments{
\item{weather}{data.frame or path to file}
}
\value{
Weather Data Frame
}
\description{
Wrangle weather data.\cr
\itemize{
\item Round ts and add hour, date, month, year (to join with obs) \cr
\item Rename city_name -> city \cr
\item Rename to outdoor_hum & outdoor_temp \cr
}
}
\examples{
wrangle_weather(filtered_weather = test_params("monkey_a")$weather)

}
