% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-full-summary.R
\name{device_excluded_kable}
\alias{device_excluded_kable}
\title{excluded_device_kable}
\usage{
device_excluded_kable(devices_for_map)
}
\arguments{
\item{devices_for_map}{from \code{get_devices_for_map()}}
}
\value{
kable and description line
}
\description{
Summary Table of Devices Excluded
}
