% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c-home-report-Ventilation.R
\name{get_ventilation_totals}
\alias{get_ventilation_totals}
\title{get_ventilation_totals}
\usage{
get_ventilation_totals(ventilation_data)
}
\arguments{
\item{ventilation_data}{data.frame, the output from \code{?ventilation_data()}}
}
\description{
Totals for Value Boxes
}
