% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b-wrangle.R
\name{memory_saver_obs}
\alias{memory_saver_obs}
\title{memory_saver_obs}
\usage{
memory_saver_obs(wrangled_obs)
}
\arguments{
\item{wrangle_obs}{from \code{wrangle_observations}}
}
\description{
Cut out unnecessary data
orig size   = 777647056  = 100\%
filter size = 388841984  =  50\%
select size = 218039344  =  28\%
}
\examples{
wrangle_obs <- memory_saver_obs(wrangle_obs)

}
