% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{monkey_knit_error}
\alias{monkey_knit_error}
\title{monkey_error}
\usage{
monkey_knit_error(err = "err message", resource = "failing r function", ...)
}
\arguments{
\item{err}{error message to report before knitr exits}

\item{resource}{the function or chunk name where it's called}
}
\description{
R Code crashed. Need to print out message to: \cr
\itemize{
\item to html report
\item console / terminal
\item render
}
}
\examples{
monkey_knit_error(err = "Error Message", resource="CHUNK_NAME")

}
