% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c-options.R
\name{monkey_palettes}
\alias{monkey_palettes}
\title{monkey_palettes : Doesn't seem to actually work...}
\usage{
monkey_palettes(devices)
}
\arguments{
\item{devices}{data.frame, the wrangled devices data as output by \code{wrangle_devices},
and further filtered by \code{remove_excluded_devices}}
}
\description{
define some color palettes and set them as R options to be globally available
}
