% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a-knit.R
\name{run_test_report}
\alias{run_test_report}
\title{Run a Test Report}
\usage{
run_test_report(
  report = "hhi",
  dummy_data = "monkey_a",
  dummy_params = "params_1.txt"
)
}
\arguments{
\item{report}{The report name}

\item{dummy_data}{The Dummy data set which can be considered as a test case}

\item{dummy_params}{The Dummy parameter set - rmd params into report}
}
\value{
an html document printed out to a file and the path of the output file is returned
}
\description{
Test Report Locally with Dummy Test Data
and Dummy Test Parameters. Uses library of test cases. Run this function first
to ensure everything is working properly and to see an example of the output.
Please note that the pameters for test reports are preset.
}
\examples{
run_test_report("home", "monkey_32",   "params_1.txt")
run_test_report("home", "monkey_a",    "params_1.txt")
run_test_report("hhi",  "org_corner2", "map.txt")

}
